# cursor/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RTR User Authentication Service",
    "description": "A comprehensive user authentication and tenant management service with JWT-based authentication, user management, and tenant settings.",
    "version": "1.0.0",
    "contact": {
      "name": "RTR Development Team",
      "email": "dev@rtr.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    },
    {
      "url": "https://api.rtr.com",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "Authentication",
      "description": "User authentication and authorization endpoints"
    },
    {
      "name": "Users",
      "description": "User management operations"
    },
    {
      "name": "Tenant Settings",
      "description": "Tenant configuration management"
    },
    {
      "name": "Health",
      "description": "Service health and status endpoints"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["Health"],
        "summary": "Health Check",
        "description": "Returns the service status and health information",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "rtr-user-auth-service: ok"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User Login",
        "description": "Authenticate a user with email and password, returns JWT token and user information",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "headers": {
              "X-Tenant-ID": {
                "description": "Tenant ID of the authenticated user",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get Current User",
        "description": "Get the current authenticated user's information",
        "operationId": "getMe",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List Users",
        "description": "Get a list of all users in the current tenant",
        "operationId": "listUsers",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create User",
        "description": "Create a new user in the current tenant",
        "operationId": "createUser",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Email already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/change-password": {
      "post": {
        "tags": ["Users"],
        "summary": "Change Password",
        "description": "Change the current user's password",
        "operationId": "changePassword",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Bad request - Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Invalid current password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tenant/settings": {
      "get": {
        "tags": ["Tenant Settings"],
        "summary": "Get Tenant Settings",
        "description": "Retrieve the current tenant's configuration settings",
        "operationId": "getTenantSettings",
        "responses": {
          "200": {
            "description": "Tenant settings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Tenant not resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Tenant Settings"],
        "summary": "Update Tenant Settings",
        "description": "Update the current tenant's configuration settings",
        "operationId": "updateTenantSettings",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tenant settings updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input data or tenant not resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from the login endpoint"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "User's password",
            "example": "password123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "Token": {
            "type": "string",
            "description": "JWT access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "ExpiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Token expiration timestamp",
            "example": "2024-01-01T12:00:00Z"
          },
          "User": {
            "$ref": "#/components/schemas/UserRead"
          }
        }
      },
      "UserRead": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "string",
            "format": "uuid",
            "description": "Unique user identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "TenantID": {
            "type": "string",
            "format": "uuid",
            "description": "Tenant identifier",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "Name": {
            "type": "string",
            "description": "User's full name",
            "example": "John Doe"
          },
          "Email": {
            "type": "string",
            "format": "email",
            "description": "User's email address",
            "example": "john.doe@example.com"
          },
          "Role": {
            "$ref": "#/components/schemas/Role"
          },
          "MustChangePassword": {
            "type": "boolean",
            "description": "Whether the user must change their password on next login",
            "example": false
          }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": ["email", "name", "role"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address",
            "example": "newuser@example.com"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "description": "User's full name",
            "example": "Jane Smith"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
      "CreateUserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserRead"
          },
          "temporary_password": {
            "type": "string",
            "description": "Temporary password for the new user",
            "example": "TempPass123!"
          }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": ["current_password", "new_password"],
        "properties": {
          "current_password": {
            "type": "string",
            "minLength": 6,
            "description": "Current password",
            "example": "oldpassword123"
          },
          "new_password": {
            "type": "string",
            "minLength": 6,
            "description": "New password",
            "example": "newpassword123"
          }
        }
      },
      "TenantSettingsRequest": {
        "type": "object",
        "required": ["config"],
        "properties": {
          "config": {
            "type": "object",
            "description": "Tenant configuration settings as key-value pairs",
            "additionalProperties": true,
            "example": {
              "theme": "dark",
              "max_users": 100,
              "features": ["analytics", "reporting"]
            }
          }
        }
      },
      "TenantSettingsResponse": {
        "type": "object",
        "properties": {
          "config": {
            "type": "object",
            "description": "Current tenant configuration settings",
            "additionalProperties": true,
            "example": {
              "theme": "dark",
              "max_users": 100,
              "features": ["analytics", "reporting"]
            }
          }
        }
      },
      "Role": {
        "type": "string",
        "enum": ["ADMIN", "HR", "INTERVIEWER", "CANDIDATE"],
        "description": "User role within the tenant",
        "example": "ADMIN"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid credentials"
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "example": "INVALID_CREDENTIALS"
          },
          "details": {
            "type": "object",
            "description": "Additional error details",
            "additionalProperties": true
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Validation error message",
            "example": "Validation failed"
          },
          "field_errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Field name that failed validation",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "description": "Validation error message for the field",
                  "example": "email is required"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}